Update Pinned Certificates (#628)
* Add pinned cert to GraphQLConstants.

* Rename BraintreeGraphQLCertificate class to PinnedCertificates.

* Move GraphQL pinned certificates from constants file into PinnedCertificates.java.

* Consolidate certificate pinning to one chain of certificates for both GraphQL and Gateway requests.

* Rename PinnedCertificates.java to TLSCertificatePinning.java.

* Remove extraneous certificates from TSLCertificatePinning constants.

* Add QA SSL Certificate Pinning test.
