--- BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeGraphQLCertificate.java
@@ -1,18 +1,0 @@
-package com.braintreepayments.api;
-
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
-
-import javax.net.ssl.SSLException;
-
-class BraintreeGraphQLCertificate {
-
-    static InputStream getCertInputStream() throws SSLException {
-        try {
-            return new ByteArrayInputStream(GraphQLConstants.CERTIFICATE.getBytes("UTF-8"));
-        } catch (UnsupportedEncodingException e) {
-            throw new SSLException(e.getMessage());
-        }
-    }
-}
